general/move_and_repeat.py:Add command delay option
Add option "--command-delay" to "move_and_repeat.py" script
to allow for a delay between updating the set point PVs values
and setting the command PVs to 1.

This is useful when the controlled IOC does not guarantee that
the device always process commands in the order they are received,
or if it is necessary some time between a motion command can be sent.